services/teams-bot/src/bot/adapters/auth_adapter.py
"""
Microsoft Teams SSO Authentication Adapter
OAuth 2.0 authentication flow using Microsoft Graph SDK for Teams SSO integration.
"""
from __future__ import annotations

import uuid
from typing import Any, Dict, Optional

import structlog
from msgraph import GraphServiceClient
from msgraph.generated.models.o_data_errors.o_data_error import ODataError

from config.settings import teams_config

# Setup structured logging
logger = structlog.get_logger(__name__)


class AuthenticationError(Exception):
    """Custom exception for authentication failures."""
    pass


class TeamsAuthAdapter:
    """
    Microsoft Teams SSO authentication adapter using Microsoft Graph SDK.
    
    Handles OAuth 2.0 authentication flow, user context extraction,
    and Microsoft Graph API interactions for Teams users.
    """
    
    def __init__(self):
        """Initialize Teams authentication adapter."""
        self.correlation_id = str(uuid.uuid4())
        
        # Initialize Microsoft Graph client
        # Note: In production, this would use proper credential management
        self.graph_client = None  # Will be initialized with proper credentials
        
        logger.info(
            "Teams auth adapter initialized",
            correlation_id=self.correlation_id,
            service="teams-bot",
            graph_base_url=teams_config.MICROSOFT_GRAPH_BASE_URL
        )
    
    async def authenticate_user_from_token(self, access_token: str) -> Dict[str, Any]:
        """
        Authenticate user using Teams SSO access token.
        
        Args:
            access_token: OAuth 2.0 access token from Teams SSO
            
        Returns:
            Dictionary containing user profile information
            
        Raises:
            AuthenticationError: If authentication fails
        """
        correlation_id = str(uuid.uuid4())
        
        try:
            logger.info(
                "Authenticating user with Teams SSO token",
                correlation_id=correlation_id,
                service="teams-bot",
                token_length=len(access_token) if access_token else 0
            )
            
            # TODO: Initialize Graph client with access token
            # This will be implemented when proper credential flow is set up
            
            # For now, return mock user profile data
            # In production, this would call Microsoft Graph API
            mock_user_profile = {
                "id": "teams-user-123",
                "displayName": "Teams User",
                "mail": "user@example.com",
                "userPrincipalName": "user@example.com",
                "tenantId": "tenant-123"
            }
            
            logger.info(
                "User authentication successful",
                correlation_id=correlation_id,
                service="teams-bot",
                user_id=mock_user_profile["id"],
                display_name=mock_user_profile["displayName"]
            )
            
            return mock_user_profile
            
        except Exception as error:
            logger.error(
                "User authentication failed",
                correlation_id=correlation_id,
                service="teams-bot",
                error=str(error),
                error_type=type(error).__name__
            )
            raise AuthenticationError(f"Authentication failed: {str(error)}")
    
    async def get_user_profile(self, user_id: str) -> Optional[Dict[str, Any]]:
        """
        Get user profile information from Microsoft Graph.
        
        Args:
            user_id: Teams user ID
            
        Returns:
            User profile dictionary or None if not found
        """
        correlation_id = str(uuid.uuid4())
        
        try:
            logger.info(
                "Fetching user profile",
                correlation_id=correlation_id,
                service="teams-bot",
                user_id=user_id
            )
            
            # TODO: Implement Microsoft Graph API call
            # This will use Graph SDK to fetch user profile
            
            # Mock response for now
            mock_profile = {
                "id": user_id,
                "displayName": "Teams User",
                "mail": "user@example.com",
                "jobTitle": "Employee",
                "department": "IT",
                "mobilePhone": None,
                "officeLocation": None
            }
            
            return mock_profile
            
        except Exception as error:
            logger.error(
                "Failed to fetch user profile",
                correlation_id=correlation_id,
                service="teams-bot",
                user_id=user_id,
                error=str(error),
                error_type=type(error).__name__
            )
            return None
    
    async def get_user_organization_context(self, user_id: str) -> Optional[Dict[str, Any]]:
        """
        Get user's organization context from Microsoft Graph.
        
        Args:
            user_id: Teams user ID
            
        Returns:
            Organization context dictionary or None if not available
        """
        correlation_id = str(uuid.uuid4())
        
        try:
            logger.info(
                "Fetching user organization context",
                correlation_id=correlation_id,
                service="teams-bot",
                user_id=user_id
            )
            
            # TODO: Implement Microsoft Graph API call
            # This would call /me/memberOf to get organization memberships
            
            # Mock response for now
            mock_org_context = {
                "tenant_id": "tenant-123",
                "organization": "Example Corp",
                "groups": ["All Company", "IT Department"],
                "roles": ["Employee"]
            }
            
            return mock_org_context
            
        except Exception as error:
            logger.error(
                "Failed to fetch organization context",
                correlation_id=correlation_id,
                service="teams-bot",
                user_id=user_id,
                error=str(error),
                error_type=type(error).__name__
            )
            return None
    
    async def validate_teams_token(self, auth_header: str) -> bool:
        """
        Validate Teams authentication token.
        
        Args:
            auth_header: Authorization header from Teams request
            
        Returns:
            True if token is valid, False otherwise
        """
        correlation_id = str(uuid.uuid4())
        
        try:
            if not auth_header or not auth_header.startswith("Bearer "):
                logger.warning(
                    "Invalid authorization header format",
                    correlation_id=correlation_id,
                    service="teams-bot",
                    has_header=bool(auth_header)
                )
                return False
            
            # Extract token from header
            token = auth_header.replace("Bearer ", "").strip()
            
            if not token:
                logger.warning(
                    "Empty authentication token",
                    correlation_id=correlation_id,
                    service="teams-bot"
                )
                return False
            
            # TODO: Implement proper JWT token validation
            # This would validate the token signature and claims
            
            logger.info(
                "Token validation performed",
                correlation_id=correlation_id,
                service="teams-bot",
                token_length=len(token)
            )
            
            return True
            
        except Exception as error:
            logger.error(
                "Token validation failed",
                correlation_id=correlation_id,
                service="teams-bot",
                error=str(error),
                error_type=type(error).__name__
            )
            return False
    
    def _create_user_friendly_error_message(self, error: Exception) -> str:
        """
        Create user-friendly error message without exposing internal details.
        
        Args:
            error: Original exception
            
        Returns:
            User-friendly error message
        """
        if isinstance(error, AuthenticationError):
            return "Authentication failed. Please try signing in again."
        elif isinstance(error, ODataError):
            return "Unable to access user information. Please check your permissions."
        else:
            return "An authentication error occurred. Please try again later."