services/teams-bot/requirements.txt
# Teams Bot Service Dependencies
aiohttp>=3.8.0

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database & ORM  
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

# Microsoft Teams Integration
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0

# HTTP Client
httpx>=0.25.0

# Configuration & Environment
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code Quality
ruff>=0.1.0